+++ WoWsShipBuilder.UI/UserControls/ShipStatsControl.axaml.cs
@@ -9,6 +9,7 @@ using WoWsShipBuilder.Core.DataUI;
 using WoWsShipBuilder.UI.ViewModels;
 using WoWsShipBuilder.UI.Views;
 using WoWsShipBuilderDataStructures;
+using static WoWsShipBuilder.UI.ViewModels.DispersionGraphViewModel;
 
 namespace WoWsShipBuilder.UI.UserControls
 {
@@ -26,27 +27,17 @@ namespace WoWsShipBuilder.UI.UserControls
 
         public void OpenDispersionGraphWindow(object sender, PointerReleasedEventArgs e)
         {
-            var dc = DataContext as ShipStatsControlViewModel;
-            var mainBattery = dc!.CurrentShipStats!.MainBatteryUI!;
-            var win = new DispersionGraphsWindow();
-            var apShellName = dc.CurrentShipStats.MainBatteryUI!.ShellData.First(x => x.Type.Equals("ap", System.StringComparison.InvariantCultureIgnoreCase)).Index;
-            var apShell = (ArtilleryShell)AppDataHelper.Instance.GetProjectile(apShellName);
-            win.DataContext = new DispersionGraphViewModel(win, mainBattery.DispersionData, (double)mainBattery.Range * 1000, dc.CurrentShipStats.Index, apShell, DispersionGraphViewModel.Tabs.Dispersion);
-            win.Show((Window)this.GetVisualRoot());
-            e.Handled = true;
+            OpenGraphsWindow(sender, e, Tabs.Dispersion);
+        }
+
+        public void OpenDispersionPlotWindow(object sender, PointerReleasedEventArgs e)
+        {
+            OpenGraphsWindow(sender, e, Tabs.Plot);
         }
 
         public void OpenBallisticGraphWindow(object sender, PointerReleasedEventArgs e)
         {
-            var dc = DataContext as ShipStatsControlViewModel;
-            var mainBattery = dc!.CurrentShipStats!.MainBatteryUI!;
-            var win = new DispersionGraphsWindow();
-            var textBlock = (TextBlock)sender;
-            var shellIndex = ((ShellUI)textBlock.DataContext!).Index;
-            var shell = AppDataHelper.Instance.GetProjectile<ArtilleryShell>(shellIndex);
-            win.DataContext = new DispersionGraphViewModel(win, mainBattery.DispersionData, (double)mainBattery.Range * 1000, dc.CurrentShipStats.Index, shell, DispersionGraphViewModel.Tabs.Ballistic);
-            win.Show((Window)this.GetVisualRoot());
-            e.Handled = true;
+            OpenGraphsWindow(sender, e, Tabs.Ballistic);
         }
 
         public void OpenTurretAnglesWindow(object sender, PointerReleasedEventArgs e)
@@ -54,6 +45,19 @@ namespace WoWsShipBuilder.UI.UserControls
             var dc = (ShipStatsControlViewModel)DataContext!;
             var win = new FiringAngleWindow();
             win.DataContext = new FiringAngleViewModel(dc.CurrentShipStats!.MainBatteryUI!.OriginalMainBatteryData);
+            win.Show((Window)this.GetVisualRoot());
+            e.Handled = true;
+        }
+
+        private void OpenGraphsWindow(object sender, PointerReleasedEventArgs e, Tabs tab)
+        {
+            var dc = DataContext as ShipStatsControlViewModel;
+            var mainBattery = dc!.CurrentShipStats!.MainBatteryUI!;
+            var win = new DispersionGraphsWindow();
+            var textBlock = (TextBlock)sender;
+            var shellIndex = ((ShellUI)textBlock.DataContext!).Index;
+            var shell = AppDataHelper.Instance.GetProjectile<ArtilleryShell>(shellIndex);
+            win.DataContext = new DispersionGraphViewModel(win, mainBattery.DispersionData, (double)mainBattery.Range * 1000, dc.CurrentShipStats.Index, shell, tab, mainBattery.Sigma);
             win.Show((Window)this.GetVisualRoot());
             e.Handled = true;
         }