+++ WoWsShipBuilder.UI/Views/StartingMenuWindow.axaml.cs
@@ -18,6 +18,19 @@ namespace WoWsShipBuilder.UI.Views
         private void InitializeComponent()
         {
             AvaloniaXamlLoader.Load(this);
+            var control = this.FindControl<Image>("Image");
+#if DEBUG
+            control.PointerReleased += Control_PointerReleased;
+#else
+            control.IsHitTestVisible = false;
+#endif
+        }
+
+        private void Control_PointerReleased(object? sender, Avalonia.Input.PointerReleasedEventArgs e)
+        {
+            var window = new TestWindow();
+            window.DataContext = new TestWindowViewModel();
+            window.Show();
         }
 
         private void LoadBuild(object? sender, Avalonia.Interactivity.RoutedEventArgs e)